requirements.txt
struct
os
sys
csv
json
umodbus
paho-mqtt
socket
datetime
time
math
signal
getopt
